requirements.txt
# Core
python-dateutil>=2.8.2
pytz>=2023.3

# Data & Math
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Technical Analysis
TA-Lib>=0.4.28
pandas>=2.0.0

# Trading APIs
python-binance>=1.0.19
MetaTrader5>=5.0.45

# Backtesting
backtrader>=1.9.78.123
matplotlib>=3.7.0

# ML Enhancement (optional)
imbalanced-learn>=0.11.0
xgboost>=2.0.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
requests>=2.31.0

# Logging & Monitoring
colorlog>=6.7.0